src/requirements.txt
fastapi
uvicorn
python-multipart  # for images
anyscale
pyyaml
'ray[serve]'
torch
diffusers
pillow